requirements.txt
openai==0.28
requests
python-dotenv
Pillow
